requirements.txt
langchain-community>=0.0.10
duckduckgo-search>=4.1.1
langgraph>=0.0.10
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.7
faiss-cpu>=1.7.4
python-pptx>=0.6.21
PyPDF2>=3.0.1
langchain
langchain_community
pydantic_ai
tiktoken